Set argv for Python client component and imported Python module
When Python is initialised as a result of a Yast module import from
another yast language binding (e.g. ruby, perl) we need to set up Argv.

Note Failure to set Argv causes some python code to fail (e.g. import
keyring)

Similarly when a Python client is invoked (via y2base) we also need to
set Argv, unlike the import scenario (where the hosting language is not
python) we ammend the arguments passed to python so that arg[0] is
pointing at the python script and not the embedding executable (e.g.
y2base)

Signed-off-by: Noel Power <noel.power@suse.com>